# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Api_projeto_Zenvia",
      "version": "v1"
    },
    "paths": {
      "/Customer": {
        "post": {
          "tags": [
            "Customer"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "Customer"
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "customerName",
              "in": "query",
              "schema": {
                "type": "string",
                "nullable": true
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/CustomerSearch"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CustomerSearch"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CustomerSearch"
                  }
                }
              }
            }
          }
        }
      },
      "/Customer/{id}": {
        "get": {
          "tags": [
            "Customer"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Customer"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Customer"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Telephone": {
          "required": [
            "areaCode",
            "customerId",
            "phoneName",
            "phoneNumber"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "customerId": {
              "type": "integer",
              "format": "int32"
            },
            "phoneName": {
              "type": "string"
            },
            "areaCode": {
              "maxLength": 2,
              "minLength": 2,
              "pattern": "^[0-9]*$",
              "type": "string"
            },
            "phoneNumber": {
              "maxLength": 9,
              "minLength": 9,
              "pattern": "^[0-9]*$",
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "Address": {
          "required": [
            "addressName",
            "customerId",
            "street"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "customerId": {
              "type": "integer",
              "format": "int32"
            },
            "addressName": {
              "type": "string"
            },
            "street": {
              "type": "string"
            },
            "number": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "complement": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "Customer": {
          "required": [
            "birthDate",
            "cpf",
            "name",
            "rg"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string"
            },
            "cpf": {
              "maxLength": 11,
              "minLength": 11,
              "pattern": "^[0-9]*$",
              "type": "string"
            },
            "rg": {
              "maxLength": 7,
              "minLength": 7,
              "pattern": "^[0-9]*$",
              "type": "string"
            },
            "telephones": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Telephone"
              },
              "nullable": true
            },
            "adresses": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Address"
              },
              "nullable": true
            },
            "birthDate": {
              "type": "string",
              "format": "date-time"
            },
            "facebook": {
              "type": "string",
              "nullable": true
            },
            "twitter": {
              "type": "string",
              "nullable": true
            },
            "instagram": {
              "type": "string",
              "nullable": true
            },
            "linkedIn": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "CustomerSearch": {
          "type": "object",
          "properties": {
            "totalCount": {
              "type": "integer",
              "format": "int32"
            },
            "customers": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Customer"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      }
    }
  }